pyproject.toml
[project]
name = "octantis"
version = "0.1.0"
description = "Intelligent infrastructure monitoring agent for EKS/K8s"
readme = "README.md"
requires-python = ">=3.12"
license = { text = "AGPL-3.0-or-later" }
classifiers = [
    "License :: OSI Approved :: GNU Affero General Public License v3 or later (AGPLv3+)",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "langgraph>=0.2.0",
    "langchain-anthropic>=0.3.0",
    "litellm>=1.83.0",
    "grpcio>=1.62.0",
    "opentelemetry-proto>=1.24.0",
    "aiohttp>=3.9.0",
    "mcp[sse]>=1.0.0",
    "langchain-mcp-adapters>=0.1.0",
    "prometheus-client>=0.21.0",
    "httpx>=0.27.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",
    "structlog>=24.4.0",
    "python-dotenv>=1.0.0",
    "pytest-cov>=7.1.0",
    "octantis-plugin-sdk",
    "pyjwt[crypto]>=2.9.0",
    "aiokafka>=0.11.0",
]

[tool.uv.sources]
octantis-plugin-sdk = { path = "packages/octantis-plugin-sdk", editable = true }

[project.scripts]
octantis = "octantis.main:main"

[project.entry-points."octantis.ingesters"]
otlp-grpc = "octantis.plugins.builtins.ingester_plugins:OTLPGrpcIngester"
otlp-http = "octantis.plugins.builtins.ingester_plugins:OTLPHttpIngester"

[project.entry-points."octantis.processors"]
trigger-filter = "octantis.plugins.builtins.trigger_filter_plugin:TriggerFilterPlugin"
fingerprint-cooldown = "octantis.plugins.builtins.cooldown_plugin:FingerprintCooldownPlugin"

[project.entry-points."octantis.notifiers"]
slack = "octantis.plugins.builtins.notifier_plugins:SlackNotifierPlugin"
discord = "octantis.plugins.builtins.notifier_plugins:DiscordNotifierPlugin"

[project.entry-points."octantis.mcp"]
grafana-mcp = "octantis.plugins.builtins.mcp_plugins:GrafanaMCPPlugin"
k8s-mcp = "octantis.plugins.builtins.mcp_plugins:K8sMCPPlugin"
docker-mcp = "octantis.plugins.builtins.mcp_plugins:DockerMCPPlugin"
aws-mcp = "octantis.plugins.builtins.mcp_plugins:AWSMCPPlugin"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/octantis"]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-mock>=3.14.0",
    "ruff>=0.11.0",
]

[tool.ruff]
target-version = "py312"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM", "RUF"]
ignore = ["E501", "UP042"]

[tool.ruff.lint.isort]
known-first-party = ["octantis"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "--cov=octantis --cov-report=term-missing --cov-fail-under=94"

[tool.coverage.run]
source = ["src/octantis"]
omit = ["*/tests/*"]

[tool.coverage.report]
fail_under = 94
